+++ src/routes/Portal.tsx
@@ -1,4 +1,5 @@
 import { useEffect, useState, useRef } from 'react';
+import { useParams, useNavigate } from 'react-router-dom';
 import zuStore from '../store';
 import Requests from '../services/Requests';
 import Date from '../services/Date';
@@ -10,37 +11,67 @@ import Title from '../components/atoms/Title';
 import Pagination from '../components/atoms/Pagination';
 import usePagination from '../hooks/usePagination';
 import About from '../components/molecules/About';
-import { ERROR } from '../services/enums';
+
+import { ENDPOINT, ERROR } from '../services/enums';
 
 function Home() {
   const openModal = zuStore((store: any) => store.openModal);
 
-  const scollToRef = useRef<HTMLDivElement | null>(null);
+  /**
+   * ROUTE LOGIC
+  */
+  const { slug, page: paginationNumber } = useParams();
+  const navigate = useNavigate();
 
-  // DATA STATE
+  const isCategoryPage = Boolean(slug);
+  const isNumber = !Number.isNaN(Number(slug));
+
+  // wrong params and redirects
+  useEffect(() => {
+    if (isNumber) navigate('/');
+  }, []);
+
+  /**
+   * STATE LOGIC
+  */
   const [queryString, setQueryString] = useState({
     after: Date.todayDate(),
     per_page: 12,
+    page: 1,
   });
 
-  // LOADING AND PAGINATION
+  const [events, setEvents] = useState([]);
+  const [totalEvents, setTotalEvents] = useState(0);
+  const [categoryName, setCategoryName] = useState('');
+  const [errorMsg, setErrorMsg] = useState('');
+
+  /**
+   * PAGINATION LOGIC
+  */
   const {
     activePage, setActive, goPrevious, goNext,
   } = usePagination();
 
-  // const [highlights, setHighlight] = useState([]);
-  const [events, setEvents] = useState([]);
-  const [totalEvents, setTotalEvents] = useState(0);
-  const [errorMsg, setErrorMsg] = useState('');
+  useEffect(() => {
+    setQueryString((state) => ({
+      ...state,
+      page: activePage,
+    }));
+  }, [activePage]);
 
+  /**
+   * DATA EVENTS LOGIC
+  */
   useEffect(() => {
     const getEvents = async () => {
       try {
-        const result = await Requests.getEvents('events', queryString);
+        const result = await Requests.getEvents(isCategoryPage ? `${ENDPOINT.CATEGORY}${slug}` : 'events', queryString);
 
         if (!result) {
           return;
         }
+
+        setCategoryName(() => result.data.name);
 
         const resultIsEmpty = result.data.posts.length === 0;
 
@@ -60,15 +91,12 @@ function Home() {
     };
 
     getEvents();
-  }, [queryString]);
-  // end
+  }, [queryString, activePage]);
 
-  useEffect(() => {
-    setQueryString((state) => ({
-      ...state,
-      page: activePage,
-    }));
-  }, [activePage]);
+  /**
+   * UI LOGIC
+  */
+  const scollToRef = useRef<HTMLDivElement | null>(null);
 
   const scrollPageUp = () => {
     if (!scollToRef.current) return;
@@ -81,12 +109,31 @@ function Home() {
   return (
     <>
       <div className="main-events my-24" ref={scollToRef}>
-        <Title>principais shows e festas no rio de janeiro</Title>
+        <Title>
+          {
+            isCategoryPage
+              ? `Melhores shows e festas em ${categoryName}` : 'Principais shows e festas no Rio de Janeiro'
+          }
+
+        </Title>
 
         <CardGrid>
           {errorMsg ? (
             <ErrorCard>
               <p className="text-2xl">{errorMsg}</p>
+              <br />
+              <p className="font-normal">
+                Que tal,
+                {' '}
+                <a
+                  className="font-bold underline hover:no-underline"
+                  href="/"
+                  target="_self"
+                  rel="noopener noreferrer"
+                >
+                  voltar para a Home?
+                </a>
+              </p>
             </ErrorCard>
           ) : null}
