+++ src/routes/index.tsx
@@ -1,30 +1,15 @@
 import { createBrowserRouter } from 'react-router-dom';
 
-import Home from './Home';
-import Category from './Category';
-// import SingleEvent from './SingleEvent';
+import Portal from './Portal';
+// import Category from './Category';
 import Error from './Error';
 
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <Home />,
-    errorElement: <Error />,
-  },
-  {
-    path: '/category/:slug',
-    element: <Category />,
-    errorElement: <Error />,
-  },
-  // {
-  //   path: "/event/:slug",
-  //   element: <SingleEvent />,
-  //   // errorElement: <Error />,
-  // },
-  // {
-  //   element: <Error />,
-  //   // errorElement: <Error />,
-  // },
-]);
+const allRoutes = ['/', '/category/:slug', '/page/:page', '/category/:slug/page/:page'].map((path) => ({
+  path,
+  element: <Portal />,
+  errorElement: <Error />,
+}));
+
+const router = createBrowserRouter(allRoutes);
 
 export default router;
